Move JAX export to separate file
