+++ pysr/__init__.py
@@ -1,3 +1,3 @@
 from .sr import pysr, get_hof, best, best_tex, best_callable, best_row
 from .feynman_problems import Problem, FeynmanProblem
-from .export import sympy2jax
+from .export_jax import sympy2jax
