+++ pysr/sr.py
@@ -13,7 +13,7 @@ import shutil
 from pathlib import Path
 from datetime import datetime
 import warnings
-from .export import sympy2jax
+from .export_jax import sympy2jax
 
 global_equation_file = 'hall_of_fame.csv'
 global_n_features = None
